# Conflicts:
#	app/src/main/java/com/example/kamteamapp/MyNavGraph.kt
#	app/src/main/java/com/example/kamteamapp/ui/chat/ConversationUiState.kt

+++ app/src/main/java/com/example/kamteamapp/ui/HistoryProgram/HistoryProgramScreen.kt
@@ -156,7 +156,7 @@ fun MyItem(
                 )
                 Spacer(Modifier.weight(1f))
                 Text(
-                    text = stringResource(id = R.string.in_stock,item.trval_day),
+                    text = stringResource(id = R.string.in_stock,item.trval_day.toInt()),
                     style = MaterialTheme.typography.Card5Content3
                 )
             }
